refactor: 사용자의 입력에 대한 검증 책임을 분리
